refactor(iri): extract the Turtle RFC 3986 resolver into fluree-vocab::iri
The Turtle parser owned the only spec-correct IRI reference resolver in the
workspace (RFC 3986 §5: fragment rules, dot-segment removal, authority/path
merging). SPARQL lowering needs the identical semantics for BASE-relative
resolution, so the resolver moves to the dependency-free fluree-vocab crate
as pub fns is_absolute_iri / resolve_iri, with unit tests covering the
W3C-relevant shapes (sibling-document refs, fragment refs, <> against a
fragmented base, dot segments, scheme-only absolutes like urn:/did:).

The Turtle parser now delegates; its resolve_iri keeps only the
Turtle-specific relative-without-base error. Behavior is unchanged — the
existing @base integration tests in fluree-graph-turtle keep guarding the
delegated implementation.
